Make device import work again
